server restructure for scalability
