--- packages/server/src/schema/resolvers/index.js
@@ -1,1 +1,0 @@
-module.exports = [require('./user')];
