--- packages/server/src/schema/typeDefs/index.js
@@ -1,13 +1,0 @@
-const { gql } = require('apollo-server');
-
-const base = gql`
-  type Query {
-    _: String
-  }
-
-  type Mutation {
-    _: String
-  }
-`;
-
-module.exports = [base, require('./user')];
